Basic function with practical user experience
Allowed user to apply image with a button click; allowed user to zoom
in/out the image with a slider
