+++ Wallidea/Wallidea/Base.lproj/Main.storyboard
@@ -11,20 +11,51 @@
     </dependencies>
     <scenes>
         <!--View Controller-->
-        <scene sceneID="tXr-a1-R10">
+        <scene sceneID="oAJ-IM-RsJ">
             <objects>
-                <viewController id="BV1-FR-VrT" customClass="ViewController" customModule="Wallidea" customModuleProvider="target" sceneMemberID="viewController">
-                    <arskView key="view" contentMode="scaleToFill" id="EO5-fy-WZq">
+                <viewController id="6Iy-LE-hKq" customClass="ARViewController" customModule="Wallidea" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="QOF-Kl-Vhz">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                    </arskView>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <arskView contentMode="scaleToFill" fixedFrame="YES" translatesAutoresizingMaskIntoConstraints="NO" id="P0I-sN-cGw">
+                                <rect key="frame" x="0.0" y="121" width="375" height="460"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                            </arskView>
+                            <navigationBar contentMode="scaleToFill" fixedFrame="YES" translatesAutoresizingMaskIntoConstraints="NO" id="pX7-hd-r3A">
+                                <rect key="frame" x="0.0" y="20" width="375" height="44"/>
+                                <autoresizingMask key="autoresizingMask" widthSizable="YES" flexibleMaxY="YES"/>
+                                <items>
+                                    <navigationItem title="AR" id="7bq-D4-iGL"/>
+                                </items>
+                            </navigationBar>
+                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="7Wz-DT-Mk6">
+                                <rect key="frame" x="144" y="637" width="86" height="30"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <state key="normal" title="Apply Image"/>
+                                <connections>
+                                    <action selector="applyImage:" destination="6Iy-LE-hKq" eventType="touchUpInside" id="RLu-rx-q2K"/>
+                                </connections>
+                            </button>
+                            <slider opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" value="0.5" minValue="0.0" maxValue="1" translatesAutoresizingMaskIntoConstraints="NO" id="ujA-e6-rd3">
+                                <rect key="frame" x="-2" y="589" width="379" height="30"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <connections>
+                                    <action selector="imageSizeSliderChanged:" destination="6Iy-LE-hKq" eventType="valueChanged" id="LvX-3K-E7U"/>
+                                </connections>
+                            </slider>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <viewLayoutGuide key="safeArea" id="5Tk-1J-t6q"/>
+                    </view>
                     <connections>
-                        <outlet property="sceneView" destination="EO5-fy-WZq" id="YuC-8q-ovz"/>
+                        <outlet property="imageSizeSlider" destination="ujA-e6-rd3" id="fgv-xF-CeQ"/>
+                        <outlet property="sceneView" destination="P0I-sN-cGw" id="RiF-Q4-9OI"/>
                     </connections>
                 </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="SZV-WD-TEh" sceneMemberID="firstResponder"/>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="QPE-Ir-zRz" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-652" y="45"/>
+            <point key="canvasLocation" x="-652" y="81.409295352323838"/>
         </scene>
         <!--Main Page-->
         <scene sceneID="65J-OD-OQL">
@@ -45,7 +76,7 @@
                                                 <state key="normal" title="ToARKit"/>
                                                 <connections>
                                                     <action selector="toARKitButton:" destination="rCD-cI-M6q" eventType="touchUpInside" id="hsz-8J-XLg"/>
-                                                    <segue destination="BV1-FR-VrT" kind="showDetail" identifier="ToARKit" id="bqW-Fm-8gO"/>
+                                                    <segue destination="6Iy-LE-hKq" kind="showDetail" identifier="ToARKit" id="mG7-Dg-8G8"/>
                                                 </connections>
                                             </button>
                                         </subviews>
